refactor: StudioCompass 에서 레인 오버레이 둘을 갈라낸다 (2,675→2,371줄)
분할 3탄(1탄 델타 미리보기 #1071, 2탄 피커·검색 #1074). 한 레인의 이웃
전체를 훑는 오버플로 목록과 위성 제자리 편집 카드(합쳐서 305줄)를
StudioLaneOverlays.tsx 로. 둘이 쓰던 카드·위성 치수(CARD/SAT)는 공유
기하 모듈로 승격 — 이제 무대 치수의 정본이 한 곳이다. 본체에서 받는
것은 타입뿐(순환 없음). 시작점 3,523줄 대비 -33%. 동작 변화 0 —
공방 단위 274 · e2e 13 · 계약 1,639 초록.
